feat(gauges): refactor type definitions and consolidate interfaces in gaugesFixture
